Several fixes

+++ src/main/java/br/com/gitanalyzer/dto/ContributorDTO.java
@@ -1,7 +1,5 @@
 package br.com.gitanalyzer.dto;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -18,6 +16,6 @@ public class ContributorDTO {
 	private String email;
 	private boolean active;
 	private int numberFilesAuthor;
-	private BigDecimal percentOfFilesAuthored;
+	private Double percentOfFilesAuthored;
 
 }
